ManageBaseData2
可维护指标组

+++ IMS2/Controllers/IndicatorGroupsController.cs
@@ -8,6 +8,8 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using IMS2.Models;
+using IMS2.ViewModels;
+using System.Data.Entity.Infrastructure;
 
 namespace IMS2.Controllers
 {
@@ -16,9 +18,17 @@ namespace IMS2.Controllers
         private ImsDbContext db = new ImsDbContext();
 
         // GET: IndicatorGroups
-        public async Task<ActionResult> Index()
+        public async Task<ActionResult> Index(IMSMessageIdEnum? message)
         {
-            return View(await db.IndicatorGroups.ToListAsync());
+            ViewBag.StatusMessage =
+                message == IMSMessageIdEnum.CreateSuccess ? "已创建新项。"
+                : message == IMSMessageIdEnum.EditdSuccess ? "已更新完成。"
+                : message == IMSMessageIdEnum.DeleteSuccess ? "已删除成功。"
+                : message == IMSMessageIdEnum.CreateError ? "创建项目出现错误。"
+                : message == IMSMessageIdEnum.EditError ? "有重名，无法更新相关信息。"
+                : message == IMSMessageIdEnum.DeleteError ? "不允许删除该项。"
+                : "";
+            return View(await db.IndicatorGroups.OrderBy(i=>i.Priority).ToListAsync());
         }
 
         // GET: IndicatorGroups/Details/5
@@ -51,10 +61,31 @@ namespace IMS2.Controllers
         {
             if (ModelState.IsValid)
             {
-                indicatorGroup.IndicatorGroupId = Guid.NewGuid();
-                db.IndicatorGroups.Add(indicatorGroup);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                try
+                {
+                    var query = await db.IndicatorGroups.Where(d => d.IndicatorGroupId == indicatorGroup.IndicatorGroupId || d.IndicatorGroupName == indicatorGroup.IndicatorGroupName)
+                              .SingleOrDefaultAsync();
+                    if (query == null)
+                    {
+                        indicatorGroup.IndicatorGroupId = System.Guid.NewGuid();
+                        db.IndicatorGroups.Add(indicatorGroup);
+                        await db.SaveChangesAsync();
+                        return RedirectToAction("Index", new { message = IMSMessageIdEnum.CreateSuccess });
+
+                    }
+                    else
+                    {
+                        return RedirectToAction("Index", new { message = IMSMessageIdEnum.CreateError });
+                    }
+
+                }
+                catch (Exception)
+                {
+
+                    ModelState.AddModelError("", "数据库中有重名，与管理员联系解决！");
+                }
+                //查找是否有要同的Guid与相同的名称
+               
             }
 
             return View(indicatorGroup);
@@ -84,9 +115,39 @@ namespace IMS2.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Entry(indicatorGroup).State = EntityState.Modified;
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                var query = await db.IndicatorGroups.Where(i => i.IndicatorGroupName == indicatorGroup.IndicatorGroupName
+                           && i.IndicatorGroupId != indicatorGroup.IndicatorGroupId).SingleOrDefaultAsync();
+                if(query != null)
+                {
+                    ModelState.AddModelError("", String.Format("不能出现同名：{0}", indicatorGroup.IndicatorGroupName));
+                }
+                else
+                {
+                    //只能更改优先级与备注信息
+                    db.Entry(indicatorGroup).State = EntityState.Modified;
+                    //client win
+                    bool saveFailed;
+                    do
+                    {
+                        saveFailed = false;
+                        try
+                        {
+                            await db.SaveChangesAsync();
+
+                        }
+                        catch (DbUpdateConcurrencyException ex)
+                        {
+                            saveFailed = true;
+
+                            // Update original values from the database 
+                            var entry = ex.Entries.Single();
+                            entry.OriginalValues.SetValues(entry.GetDatabaseValues());
+                        }
+
+                    } while (saveFailed);
+
+                    return RedirectToAction("Index", new { message = IMSMessageIdEnum.EditdSuccess });
+                }
             }
             return View(indicatorGroup);
         }
@@ -112,9 +173,34 @@ namespace IMS2.Controllers
         public async Task<ActionResult> DeleteConfirmed(Guid id)
         {
             IndicatorGroup indicatorGroup = await db.IndicatorGroups.FindAsync(id);
-            db.IndicatorGroups.Remove(indicatorGroup);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            if(indicatorGroup.IndicatorGroupMapIndicators.Count <= 0 &&
+                indicatorGroup.DepartmentCategoryMapIndicatorGroups.Count <= 0)
+            {
+                db.IndicatorGroups.Remove(indicatorGroup);
+                //client win
+                bool saveFailed;
+                do
+                {
+                    saveFailed = false;
+                    try
+                    {
+                        await db.SaveChangesAsync();
+
+                    }
+                    catch (DbUpdateConcurrencyException ex)
+                    {
+                        saveFailed = true;
+
+                        // Update original values from the database 
+                        var entry = ex.Entries.Single();
+                        entry.OriginalValues.SetValues(entry.GetDatabaseValues());
+                    }
+
+                } while (saveFailed);
+                return RedirectToAction("Index", new { message = IMSMessageIdEnum.DeleteSuccess });
+            }
+            
+            return RedirectToAction("Index", new { message = IMSMessageIdEnum.DeleteError });
         }
 
         protected override void Dispose(bool disposing)
